clean up of planner course formatter
